update: updating cart items  and removing items from cart

+++ app/controllers/cart/delete.cart.js
@@ -9,7 +9,7 @@ export class DeleteCartController {
             if (!req.params.cartItemId) {
                 return res.ApiResponse.error(400, 'The cart ID is required');
             }
-            const {success, cart, error } = await this.deleteCartService.deleteCart(req.params.cartItemId);
+            const {success, cart, error } = await this.deleteCartService.deleteCart(req.user.userId, req.params.cartItemId, req.model);
             if (!success) {
                 return res.ApiResponse.error(400, error);
             }
